courses/atf_cloud_interactive/agents/improve_engagement_agent_tmp20260305_134949/requirements.txt
google-cloud-aiplatform==1.139.0
google-adk[a2a]==1.26.0
google-auth==2.48.0
google-genai==1.65.0
python-dotenv==1.2.1
opentelemetry-exporter-gcp-trace
opentelemetry-sdk
opentelemetry-exporter-gcp-logging
opentelemetry-instrumentation-google-genai